Modification of Coding userController and authController...

+++ src/controllers/userController.js
@@ -1,34 +1,42 @@
+const { default: mongoose } = require('mongoose');
 const User = require('../models/userModel')
 
 const getAllUsersController = async (req, res) => {
-    const { id } = req.params;
     try {
-        const userData = await User.find(id)
-        if (userData) {
-            res.status(200).json({ success: true, message: "All User data displayed...", users: userData })
-        } else {
-            res.status(404).json({ success: false, message: "User data not found..." })
-        }
-
+        const userData = await User.find({}).select("-password")
+        
+        return res.status(200).json({
+            success: true,
+            message: "All User data displayed...",
+            users: userData
+        })
+        
     } catch (error) {
-        console.log('gell all data error:...', error)
+        console.log('Get all users error:...', error)
         return res.status(500).json({ success: false, message: 'Server error...' });
     }
 };
 
 const singleUserDataController = async (req, res) => {
-    const { id } = req.params;
     try {
-        const userData = await User.findById(id)
+        const { id } = req.params;
 
-        if (userData) {
-            res.status(200).json({ success: true, message: `user: ${userData.email} data`, Data: userData })
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({
+                success: false,
+                message: "Invalid ID"
+            })
+        };
+        const userData = await User.findById(id).select("-password")
+
+        if (!userData) {
+            return res.status(404).json({ success: false, message: "User not found..." })
         } else {
-            res.status(404).json({ success: false, message: "User data not found..." })
+            return res.status(200).json({ success: true, message: `user: ${userData.email}`, user: userData })
         }
 
     } catch (error) {
-        console.log('gell Single User data error:...', error)
+        console.log('get Single User data error:...', error)
         return res.status(500).json({ success: false, message: 'Server error...' });
     }
 }
@@ -36,15 +44,23 @@ const singleUserDataController = async (req, res) => {
 const deleteDataController = async (req, res) => {
     try {
         const { id } = req.params
+        
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({
+                success: false,
+                message: "Invalid ID"
+            })
+        };
+
         const userData = await User.findByIdAndDelete(id)
         if (!userData) {
-            return res.status(404).json({ success: false, message: 'User data not found...' })
+            return res.status(404).json({ success: false, message: 'User not found...' })
         } else {
-            return res.status(200).json({ success: true, message: 'User data deleted...' })
+            return res.status(200).json({ success: true, message: 'User deleted...' })
         }
 
     } catch (error) {
-        console.log('Delete data error:...', error)
+        console.log('Delete error:...', error)
         return res.status(500).json({ success: false, message: 'Server error...' });
     }
 
@@ -53,16 +69,22 @@ const deleteDataController = async (req, res) => {
 const updateUserDataController = async (req, res) => {
     try {
         const { id } = req.params;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({
+                success: false,
+                message: "Invalid ID"
+            })
+        };
         const userData = await User.findByIdAndUpdate(id, req.body, { returnDocument: 'after' })
 
         if (!userData) {
-            return res.status(404).json({ success: false, message: 'User data not found...' })
+            return res.status(404).json({ success: false, message: 'User not found...' })
         } else {
-            return res.status(200).json({ success: true, message: 'User data Updated...' })
+            return res.status(200).json({ success: true, message: 'User Updated...' })
         }
 
     } catch (error) {
-        console.log('Update data error:...', error)
+        console.log('Update error:...', error)
         return res.status(500).json({ success: false, message: 'Server error...' });
     }
 }
